feat: Implementation of mocked data through the CSV data Data is loaded into duckdb and data functions are implemented as SQL over duckdb
Zed provided the starting implementation, which was then modified and tweaked.

Prompt:

Refactor data.ts so that the data is coming directly from the CSVs Sunshine 2025 28.05.2025.csv (representing sunshine data from 2025), Sunshine 2024 28.05.2025.csv (data from 2024), and peer-groups.csv.

The script load-sunshine-data.sql can be seen as an example of how the csv data is "parsed" into SQL tables.

Data should be first loaded in duckdb and new tables should be created so that we are able to access the data easily.

Create indexes if necessary

The mock API should only be functions that return data. In the end we should be able to return the following data (see data.ts)

- SunshineCostsAndTariffsData
- SunshinePowerStabilityData
- SunshineOperationalStandardsData

Both Sunshine csv should be loaded into duckdb (use @duckdb/node-api) and then SQL should be written to access & filter the data

In the end, the functions fetchOperatorCostsAndTariffsData, fetchPowerStability, fetchOperationalStandards should fetch their data directly from duckdb