fix: Operator was not showing on the map
This was due to hardcoded category in the operator municipalities fetching
function. Since we do not have geographical information in the sunshine data,
we need to somehow fetch the links between operator and municipalities from
the energy price cube (which contains municipality information). Previously,
we were filtering the energy price cube observations on category C2. Now
we filter on category for sunshine indicator that are category based, otherwise
we do not put a category filter